fix some constants
EPSILON = 2^-10, its biased exp is 5, so bits are 5<<10 = 0x1400.
10^5 <= MAX < 10^6, so MAX_10_EXP is 5.
10^-5 <= MIN < 10^-4, so MIN_10_EXP is -4.

MAX_EXP and MIN_EXP are 16, -13, because like C's float.h, std::f32
assumes the significand is in the range 0.5 <= x < 1.0. (The text of
the IEEE-754 standard uses 1.0 <= x < 2.0, such that its maximum and
minimum exponent values are 15 and -14.)

f32::NAN::to_bits() gives 0x7FC0_0000, so its equivalent is 0x7E00.