ties round to even when converting f16 to f32 and f16 to f64
In conversions from `f32` or `f64` to `f16`, instead of rounding to
nearest with ties rounded away from zero, now ties are rounded to
even. This can be achieved using only one extra mask check for normal
numbers, using a method explained in an inline comment. (For
subnormals, `round_bit` is not a constant so it is slightly more
expensive, but normals should be the common case.)

Some redundant checks for zero/subnormals were removed as they are
perfectly handled by the underflow condition.

Also, when converting NaNs, the most significant 10 bits of the
mantissa are retained, then the most significant bit of the mantissa
is set. This is similar to what happens when converting an `f32` NaN
to `f64`, where the most significant 23 bits of the NaN mantissa are
retained, then the most significant bit of the mantissa is set.