Add `#[repr(transparent)]` to `f16`/`bf16` types
Currently this crate wraps an `u16` using a tuple struct into a new
type. This is fine, as long the memory representation of this new type
is irrelevant (as it is in safe Rust code).
Unfortunately, there is at least one function, that reinterprets an
`u16` into a `f16` using casting: `HalfBitsVecExt::reinterpret_into()`¹.
At this point, the memory representation of the type is important. Rust
does not guarantee, that the wrapping tuple-struct has the same repre-
sentation as the wrapped type. This possibly leads to undefined behav-
ior!
At this point the `#[repr(transparent)]` attribute is helpful: it serves
exactly that purpose to guarantee the same memory layout according to
the nomicon². Using this attribute on the `f16` and `bf16` structs, the
pointer casting remains valid and removes the UB.

All previous tests still pass. I have not added a new test case, as I
don't know a way to test the presence of attributes or the correctness
of the codegen.

¹: https://docs.rs/half/1.4.0/src/half/vec.rs.html#215
²: https://doc.rust-lang.org/nomicon/other-reprs.html#reprtransparent