Fix as_ptr() -> as_mut_ptr() finding from Miri.
I'm adding a dependency on half to a large project and wanted to give it a quick
check with miri. The miri stacked borrows checker had an issue with using a non
mutable pointer in the mutable reinterpret cast. Its not clear to me if this
was really indicative of a problem, but the version miri is happy with does look
better to me.

(note there were two other miri failures but they look like a problem
with miri itself - https://github.com/rust-lang/rust/issues/69532)